feat: 코인 출금 API 구현

Closes S12P21A604-59

See merge request s12-fintech-finance-sub1/S12P21A604!229

+++ rabbit-server/src/main/java/com/rabbit/bankApi/service/BankService.java
@@ -56,7 +56,7 @@ public class BankService {
      * @param userId 유저 ID
      * @param balance 금액
      */
-    public void deposit(int userId, int balance) {
+    public void deposit(int userId, long balance) {
         // 1. userId로 이메일 조회
         String email = userRepository.findById(userId)
                 .orElseThrow(() -> new BusinessException(ErrorCode.RESOURCE_NOT_FOUND, "존재하지 않는 사용자입니다."))
